pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
requests = "^2.28.0"
Flask = "^2.1.2"
qrcode = "^7.3.1"
replit = "^3.2.4"
cryptocode = "^0.1"
Pillow = "^9.1.1"
bs4 = "^0.0.1"
config = "^0.5.1"
uptimes = "^1.0.0"
future = "^0.18.2"
baekjoon = "^1.0.8"
"discord.py" = "^1.7.3"
discord-pretty-help = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"